test out of sync measurements (fails)
